remove old folder tree file
