+++ src/replays/folderTreeService.ts
@@ -1,14 +1,9 @@
+import * as fs from "fs-extra";
 import path from "path";
 
-import { generateSubFolderTree } from "./folderTree";
 import type { FolderResult } from "./types";
 
-export interface IFolderTreeService {
-  init(folders: readonly string[]): readonly FolderResult[];
-  select(folder: string): Promise<readonly FolderResult[]>;
-}
-
-export class FolderTreeService implements IFolderTreeService {
+export class FolderTreeService {
   public tree: readonly FolderResult[] = [];
 
   public init(rootFolders: readonly string[]): readonly FolderResult[] {
@@ -38,4 +33,39 @@ export class FolderTreeService implements IFolderTreeService {
 
     return this._findChild(folder, res.subdirectories);
   }
+}
+
+/**
+ * Returns the tree structure of a folder
+ * @param folder Details including name, and subdirectories
+ * @param childrenToExpand The name of the subdirectories to expand.
+ */
+export async function generateSubFolderTree(folder: string): Promise<FolderResult[]> {
+  console.log(`generating subfolder tree for folder: ${folder}`);
+  // Only generate the tree for a single level
+  let results: fs.Dirent[] = [];
+  try {
+    // The directory we're expanding might not actually exist.
+    results = await fs.readdir(folder, { withFileTypes: true });
+  } catch (err) {
+    // If it doesn't exist, just return an empty list.
+    console.warn(err);
+    return [];
+  }
+
+  const subdirectories = results
+    .filter((dirent) => {
+      return dirent.isDirectory();
+    })
+    .map(
+      async (dirent): Promise<FolderResult> => {
+        return {
+          name: dirent.name,
+          fullPath: path.join(folder, dirent.name),
+          subdirectories: [],
+        };
+      },
+    );
+
+  return Promise.all(subdirectories);
 }
